tweaks and media

+++ src/app/manage-users/manage-users.component.ts
@@ -4,6 +4,7 @@ import { User } from '../user/user.model';
 import { BehaviorSubject, Subscription } from 'rxjs';
 import { UserService } from '../user/user.service';
 import { DataService } from '../shared/data.service';
+import Swal from 'sweetalert2';
 
 @Component({
     selector: 'app-manage-users',
@@ -40,22 +41,75 @@ export class ManageUsersComponent implements OnInit {
     }
 
     makeAdmin(user: User) {
-        user.admin = 1;
-        this.data.editUser(user);
+        Swal.fire({
+            title: "Change role?",
+            text: `You will give admin permissions to ${user.username}.`,
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, change role!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                user.admin = 1;
+                this.data.editUser(user);
+                Swal.fire({
+                    title: "Success!",
+                    text: `User ${user.username} is now admin!`,
+                    icon: "success"
+                });
+            }
+        });
     }
 
     makeUser(user: User) {
-        user.admin = 0;
-        this.data.editUser(user);
+        Swal.fire({
+            title: "Change role?",
+            text: `You will take admin permissions from ${user.username}.`,
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, change role!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                user.admin = 0;
+                this.data.editUser(user);
+                Swal.fire({
+                    title: "Success!",
+                    text: `User ${user.username} is now user!`,
+                    icon: "success"
+                });
+            }
+        });
     }
 
     deleteUser(user: User) {
-        this.users.forEach((element, index) => {
-            if (element === user)
-                this.users.splice(index, 1);
+        let username = user.username
+        Swal.fire({
+            title: "Delete user?",
+            text: `Do you want to delete user ${username}.`,
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete user!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                this.users.forEach((element, index) => {
+                    if (element === user)
+                        this.users.splice(index, 1);
+                });
+                console.log(user.userID)
+                this.auth.deleteUser(user.userID);
+                Swal.fire({
+                    title: "Success!",
+                    text: `User ${username} is deleted!`,
+                    icon: "success"
+                });
+            }
         });
-        console.log(user.userID)
-        this.auth.deleteUser(user.userID);
+
     }
 
 }
